ViewModel Modification
