+++ client/android/application/app/src/main/java/com/example/myapplication/Attempt/AttemptRepository.kt
@@ -2,15 +2,16 @@ package com.example.myapplication.Attempt
 
 import com.example.myapplication.network.Network
 import com.example.myapplication.network.TopicApi
+import retrofit2.Callback
 
-class AttemptRepository(private val callback: retrofit2.Callback<ResponseStart>) {
+class AttemptRepository(private val callback: retrofit2.Callback<Any>) {
 
     fun getQuestionsData(token: String, postStart: PostStart) {
         val apiClient = Network.getInstance().create(TopicApi::class.java)
 
         val call = apiClient.StartAttempt(token, postStart)
         call.enqueue(
-            callback
+            callback as Callback<ResponseStart>
         )
     }
