+++ client/android/application/app/src/main/java/com/example/myapplication/Attempt/QuestionsRepo.kt
@@ -4,7 +4,7 @@ import com.example.myapplication.network.Network
 import com.example.myapplication.network.TopicApi
 import retrofit2.Callback
 
-class QuestionsRepo(private val callback: Callback<ResponseQuestionData>) {
+class QuestionsRepo(private val callback: Callback<Any>) {
 
     fun getQuestionsData(
         token: String,
@@ -15,7 +15,7 @@ class QuestionsRepo(private val callback: Callback<ResponseQuestionData>) {
         val api = Network.getInstance().create(TopicApi::class.java)
 
         val call = api.getQuestionData(token, attemptId, submissionId, questionId)
-        call.enqueue(callback)
+        call.enqueue(callback as Callback<ResponseQuestionData>)
 
     }
 }
