+++ client/android/application/app/src/main/java/com/example/myapplication/HomeActivity.kt
@@ -7,7 +7,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProviders
 import com.example.myapplication.Attempt.*
 import com.example.myapplication.viewModel.AttemptViewModel
-import com.example.myapplication.viewModel.QuestionsDataViewModel
 import kotlinx.android.synthetic.main.activity_home.*
 
 class HomeActivity : AppCompatActivity() {
@@ -16,8 +15,6 @@ class HomeActivity : AppCompatActivity() {
     private lateinit var attemptViewModel: AttemptViewModel
     private lateinit var dataQuestions: DataQuestions
 
-    private lateinit var questionsDataViewModel: QuestionsDataViewModel
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -25,7 +22,7 @@ class HomeActivity : AppCompatActivity() {
         topicID = intent.getStringExtra("topicId").toString()
         token = intent.getStringExtra("token").toString()
         attemptViewModel = ViewModelProviders.of(this).get(AttemptViewModel::class.java)
-        questionsDataViewModel = ViewModelProviders.of(this).get(QuestionsDataViewModel::class.java)
+
         observeLiveData()
         observeQuestions()
         val postStart = PostStart(5, topicID)
@@ -33,8 +30,8 @@ class HomeActivity : AppCompatActivity() {
     }
 
     private fun observeQuestions() {
-        questionsDataViewModel.questionsLiveData.observe(this,{
-            when(it){
+        attemptViewModel.questionsLiveData.observe(this, {
+            when (it) {
                 is QuestionsUIModel.Success -> {
                     questionText.text = it.questionData.data?.statement
                     optionOne.text = it.questionData.data?.options?.get(0)?.text.toString()
@@ -52,12 +49,13 @@ class HomeActivity : AppCompatActivity() {
             when (it) {
                 is AttemptUIModel.Success -> {
                     Log.d("Attempt", it.startQuiz.data?.attemptId.toString())
-                    val dataStart : DataStart? = it.startQuiz.data
+                    val dataStart: DataStart? = it.startQuiz.data
                     if (dataStart != null) {
                         dataStart.questions?.get(0)?.let { it1 ->
                             dataStart.attemptId?.let { it2 ->
                                 dataStart.submissionId?.let { it3 ->
-                                    questionsDataViewModel.callQuestionsData(token, it2, it3,
+                                    attemptViewModel.callQuestionsData(
+                                        token, it2, it3,
                                         it1
                                     )
                                 }
